// src/types/settingParameter.ts
export type SettingParameter = {
  setting_id: string;
  client_id: string;
  client_secret: string;
  matter_id: string;
  jp_folder_id: string;
  sg_folder_id: string;
  vn_folder_id: string;
  target_share_drive_id: string;
  jp_row_data_folder_id: string;
  sg_row_data_folder_id: string;
  vn_row_data_folder_id: string;
};